etl/requirements.txt
psycopg2-binary==2.9.5
requests==2.25.1
pydantic==1.8.2
python-dotenv==0.19.2
elasticsearch==8.5.3
redis==4.4.0